feat: implement client feedback round 2 — 11 items
- Filter practitioner queue by logged-in doctor (item 1)
- Remove pre-consultation validation workflow entirely (item 2)
- Add schedule entry deletion with confirmation (item 3)
- Fix zone pre-population in edit dialog (item 4)
- Fix absence tracking with live updates and query invalidation (item 5)
- Show "Seance N" instead of "Seance N/Total" (item 8)
- Rework packs: months instead of days, unlimited zones/sessions, remove gold type (item 10)
- Add CSV export on agenda, absences, pre-consultations for all roles (item 11)
- Migration 022: add duree_mois, zones_illimitees, seances_illimitees to packs

+++ backend/src/schemas/paiement.py
@@ -11,7 +11,7 @@ class PaiementCreate(AppBaseModel):
     """Paiement creation schema."""
 
     patient_id: str
-    montant: int = Field(ge=0)
+    montant: float = Field(ge=0)
     type: str = Field(pattern=r"^(encaissement|prise_en_charge|hors_carte)$")
     subscription_id: str | None = None
     session_id: str | None = None
